requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0